+++ processor/src/main/java/org/realityforge/arez/processor/GeneratorUtil.java
@@ -17,6 +17,8 @@ final class GeneratorUtil
   static final String FIELD_PREFIX = "$$arez$$_";
   static final String CAUGHT_THROWABLE_NAME = FIELD_PREFIX + "e";
   static final String DISPOSED_FIELD_NAME = FIELD_PREFIX + "disposed";
+  static final String ON_DISPOSE_FIELD_NAME = FIELD_PREFIX + "onDispose";
+  static final String SET_ON_DISPOSE_METHOD_NAME = FIELD_PREFIX + "setOnDispose";
   static final String ID_FIELD_NAME = FIELD_PREFIX + "id";
   static final String NAME_METHOD_NAME = FIELD_PREFIX + "name";
   static final String NEXT_ID_FIELD_NAME = FIELD_PREFIX + "nextId";