+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithMultipleCtors.java
@@ -17,6 +17,8 @@ public final class Arez_RepositoryWithMultipleCtors extends RepositoryWithMultip
 
   private boolean $$arez$$_disposed;
 
+  private OnDispose $$arez$$_onDispose;
+
   @Nonnull
   private final ArezContext $$arez$$_context;
 
@@ -61,8 +63,16 @@ public final class Arez_RepositoryWithMultipleCtors extends RepositoryWithMultip
   public void dispose() {
     if ( !isDisposed() ) {
       $$arez$$_disposed = true;
+      if ( null != $$arez$$_onDispose ) {
+        $$arez$$_onDispose.onDispose( this );
+        $$arez$$_onDispose = null;
+      }
       $$arez$$_name.dispose();
     }
+  }
+
+  void $$arez$$_setOnDispose(OnDispose onDispose) {
+    $$arez$$_onDispose = onDispose;
   }
 
   @Nonnull
@@ -106,5 +116,10 @@ public final class Arez_RepositoryWithMultipleCtors extends RepositoryWithMultip
     } else {
       return super.toString();
     }
+  }
+
+  @FunctionalInterface
+  interface OnDispose {
+    void onDispose(Arez_RepositoryWithMultipleCtors entity);
   }
 }