+++ processor/src/test/resources/expected/com/example/repository/NestedModel$Arez_BasicActionModel.java
@@ -15,6 +15,8 @@ public final class NestedModel$Arez_BasicActionModel extends NestedModel.BasicAc
 
   private boolean $$arez$$_disposed;
 
+  private OnDispose $$arez$$_onDispose;
+
   @Nonnull
   private final ArezContext $$arez$$_context;
 
@@ -41,7 +43,15 @@ public final class NestedModel$Arez_BasicActionModel extends NestedModel.BasicAc
   public void dispose() {
     if ( !isDisposed() ) {
       $$arez$$_disposed = true;
+      if ( null != $$arez$$_onDispose ) {
+        $$arez$$_onDispose.onDispose( this );
+        $$arez$$_onDispose = null;
+      }
     }
+  }
+
+  void $$arez$$_setOnDispose(OnDispose onDispose) {
+    $$arez$$_onDispose = onDispose;
   }
 
   @Override
@@ -84,5 +94,10 @@ public final class NestedModel$Arez_BasicActionModel extends NestedModel.BasicAc
     } else {
       return super.toString();
     }
+  }
+
+  @FunctionalInterface
+  interface OnDispose {
+    void onDispose(NestedModel$Arez_BasicActionModel entity);
   }
 }